+++ src/Console/Command.php
@@ -38,6 +38,11 @@ class Command extends BaseCommand
     {
         $io->out(__d('database_backup', 'Connection: {0}', $this->getConnection()->config()['name']));
         $io->out(__d('database_backup', 'Driver: {0}', $this->getDriverName()));
+
+        if ($args->getOption('timeout')) {
+            $io->verbose(__d('database_backup', 'Timeout for shell commands: {0} seconds', $args->getOption('timeout')));
+        }
+
         $io->hr();
     }
 }